third_party/blink/web_tests/platform/generic/fast/events/message-port-gc-closed-expected.txt

MessagePorts (and their execution contexts) should be collected by GC once close() has been called on them

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS iframeDocGC.wasCollected is true
PASS successfullyParsed is true

TEST COMPLETE
